move helper methods from PackageInstaller
