+++ sampleapp/src/main/kotlin/io/github/solrudev/simpleinstaller/sampleapp/viewmodels/MainViewModel.kt
@@ -6,12 +6,12 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import io.github.solrudev.simpleinstaller.PackageInstaller
 import io.github.solrudev.simpleinstaller.data.InstallResult
+import io.github.solrudev.simpleinstaller.installPackage
 import io.github.solrudev.simpleinstaller.sampleapp.R
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
 class MainViewModel : ViewModel() {
